requirements.txt
Django==3.2.5
djangorestframework==3.12.4
Pillow
setuptools
django-crispy-forms
crispy-bootstrap4
requests